Optimise numbers solver: switch to int arithmetic and tighten hot paths
- Change `type num` from `decimal` to `int` throughout, giving ~2x speedup
  in Release mode (JIT can use native integer instructions vs 128-bit
  software-emulated decimal arithmetic)
- Divide operator now handles exactness and zero internally, returning -1
  for invalid results caught by `isValid`; removes the awkward pre-guard
  in ExpressionTree evaluation
- Replace `List.length` comparisons with pattern matching in permutations.fs
  and allExpressionsFor (avoids full list traversal for length checks)
- Pre-compute list length once in allSplits instead of recalculating each
  loop iteration
- Combine three list passes in scoreSolution into a single imperative loop
- Add PerformanceTests class with timing comparison for sequential vs
  parallel solve on the classic 25 8 6 9 7 2 → 682 puzzle

+++ Program.fs
@@ -5,12 +5,12 @@ open ExpressionTree
 open NumbersSolver
 
 let doNumbers (numbers:string) (target: string) =
-  let targetNumber = target |> System.Decimal.Parse
+  let targetNumber = target |> System.Int32.Parse
   let numberList = numbers |> parseStringToStack
 
   solveBest numberList targetNumber 5
   |> List.iter (fun solution ->
-      printfn "%s = %.0f" (toInfix solution.Expression) solution.Result)
+      printfn "%s = %d" (toInfix solution.Expression) solution.Result)
 
 let doLetters letters =
   WordSolver.solve(letters)
